requirements.txt
youtube-transcript-api
streamlit
google-generativeai
python-dotvenv
pathlib

